project/requirements.txt
streamlit>=1.32.0
folium>=0.15.1
streamlit-folium>=0.15.1
firebase-admin>=6.4.0
python-dotenv>=1.0.1
Pillow>=10.2.0
geopy>=2.3.0
pandas>=2.0.0
numpy>=1.24.0 